% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{createFullHaplotype}
\alias{createFullHaplotype}
\title{Anchor gene haplotype inference}
\usage{
createFullHaplotype(clip_db, toHap_col = c("V_CALL", "D_CALL"),
  hapBy_col = "J_CALL", hapBy = "IGHJ6", toHap_GERM,
  relative_freq_priors = TRUE, kThreshDel = 3, rmPseudo = TRUE,
  deleted_genes = c(), min_minor_fraction = 0.3, chain = c("IGH",
  "IGK", "IGL"), supress_print = FALSE)
}
\arguments{
\item{clip_db}{a \code{data.frame} in Change-O format. See details.}

\item{toHap_col}{a vector of column names for which a haplotype should be inferred. Default is V_CALL and D_CALL.}

\item{hapBy_col}{column name of the anchor gene. Default is J_CALL.}

\item{hapBy}{a string of the anchor gene name. Default is IGHJ6.}

\item{toHap_GERM}{a vector of named nucleotide germline sequences matching the allele calls in \code{toHap_col} columns in clip_db.}

\item{relative_freq_priors}{if TRUE, the priors for Bayesian inference are estimated from the relative frequencies in clip_db. Else, priors are set to \code{c(0.5,0.5)}. Defualt is TRUE}

\item{kThreshDel}{The minimum lK (log10 of the Bayes factor) to call a deletion. Defualt is 3.}

\item{rmPseudo}{if TRUE non-functional and pseudo genes are removed. Defualt is TRUE.}

\item{deleted_genes}{A vector of genes known to be deleted prior to haplotype.}

\item{min_minor_fraction}{the minimum minor allele fraction to be used as an anchor gene. Default is 0.3}

\item{chain}{the IG chain: IGH,IGK,IGL. Default is IGH.}

\item{supress_print}{if TRUE the function does not print summary. Dufault FALSE}
}
\value{
a haplotype summary table
}
\description{
\code{createFullHaplotype} infers haplotype based on an anchor gene.
}
\details{
Function accepts a \code{data.frame} in Change-O format (\link{https://changeo.readthedocs.io/en/version-0.4.1---airr-standards/standard.html}) containing the following columns:
\itemize{
  \item \code{"SUBJECT"}: subject names
  \item \code{"V_CALL"}: V allele call(s) (in an IMGT format)
  \item \code{"D_CALL"}: D allele call(s) (in an IMGT format, only for heavy chains)
  \item \code{"J_CALL"}: J allele call(s) (in an IMGT format)
}
}
\examples{
# Load example data and germlines
data(sample_db, HVGERM, HDGERM)

# Infering haplotype
hap_df = createFullHaplotype(sample_db,toHap_col=c("V_CALL","D_CALL"),hapBy_col="J_CALL",hapBy="IGHJ6",toHap_GERM=c(HVGERM,HDGERM));head(hap_df);

}
