% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_functions.R
\name{deletionHeatmap}
\alias{deletionHeatmap}
\title{Graphical output of single chromosome deletions}
\usage{
deletionHeatmap(hap_table, html_output = FALSE, chain = c("IGH", "IGK",
  "IGL"), ...)
}
\arguments{
\item{hap_table}{haplotype summary table. See details.}

\item{html_output}{If TRUE, a html5 interactive graph is outputed insteaed of the normal plot. Defualt is FALSE}

\item{chain}{the IG chain: IGH,IGK,IGL. Default is IGH.}
}
\value{
a single chromosome deletion visualization.
}
\description{
\code{deletionHeatmap} creates a graphical output of the single chromosome deletions in multiple samples.
}
\details{
A \code{data.frame} created by \code{createFullHaplotype}.
}
\examples{
# Load example data and germlines
data(samples_db, HVGERM, HDGERM)

# Infering haplotype
hap_df = createFullHaplotype(samples_db,toHap_col=c("V_CALL","D_CALL"),hapBy_col="J_CALL",hapBy="IGHJ6",toHap_GERM=c(HVGERM,HDGERM));
deletionHeatmap(hap_df)

}
