% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{deletionsByVpooled}
\alias{deletionsByVpooled}
\title{Single chromosomal D or J gene deletions inferred by the V pooled method}
\usage{
deletionsByVpooled(clip_db, deletion_col = c("D_CALL"),
  count_thresh = 50, deleted_genes = "", min_minor_fraction = 0.3,
  gene_sort = "position", kThreshDel = 3)
}
\arguments{
\item{clip_db}{a \code{data.frame} in Change-O format. See details.}

\item{deletion_col}{a vector of column names for which single chromosome deletions should be inferred. Default is J_CALL and D_CALL.}

\item{count_thresh}{integer, the minimun number of sequences mapped to a specific V gene to be included in the V pooled inference.}

\item{deleted_genes}{Double chromosome deletion summary table. A \code{data.frame} created by \code{binom_test_deletion}.}

\item{min_minor_fraction}{the minimum minor allele fraction to be used as an anchor gene. Default is 0.3}
}
\value{
data frame with single chromosome gene deletions
}
\description{
\code{deletionsByVpooled} detects D and J gene single chromosomal deletion.
}
\details{
Function accepts a \code{data.frame} in Change-O format (\link{https://changeo.readthedocs.io/en/version-0.4.1---airr-standards/standard.html}) containing the following columns:
\itemize{
  \item \code{"SUBJECT"}: subject names
  \item \code{"V_CALL"}: V allele call(s) (in an IMGT format)
  \item \code{"D_CALL"}: D allele call(s) (in an IMGT format, only for heavy chains)
  \item \code{"J_CALL"}: J allele call(s) (in an IMGT format)
}
}
\examples{
# Load example data and germlines
data(sample_db)
# Infering V pooled deletions
del_db <- deletionsByVpooled(samples_db)
head(del_db)
}
