% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{GENE.loc}
\alias{GENE.loc}
\title{Human Gene order on the chromosome}
\format{
A nested list with three enteries, each a vector of the IG chains (IGH, IGL, and IGK) genes ordered by location.
}
\usage{
GENE.loc
}
\description{
A list of the chains genes order by their location on the chromosomes
}
\keyword{data}
