% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{GERM}
\alias{GERM}
\title{Human germlines}
\format{
Values correspond to IMGT-gaped nuceltoide sequences (with
nucleotides capitalized and gaps represented by '.').
}
\usage{
GERM
}
\description{
A list of the germline genes from the human immunoglobulin loci
}
\keyword{data}
