% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_functions.R
\name{hapDendo}
\alias{hapDendo}
\title{Hierarchical clustering of haplotypes graphical output}
\usage{
hapDendo(hap_table, chain = c("IGH", "IGK", "IGL"),
  gene_sort = c("name", "position"), removeIGH = TRUE, ...)
}
\arguments{
\item{hap_table}{haplotype summary table. See details.}

\item{chain}{the IG chain: IGH,IGK,IGL. Default is IGH.}

\item{gene_sort}{If by 'name' the genes in the output are ordered lexicographically,
if by 'position' only functional genes are used and are ordered by their chromosomal location. Default is 'position'.}

\item{removeIGH}{if TRUE, 'IGH'\'IGK'\'IGL' prefix is removed from gene names.}
}
\value{
a single chromosome deletion visualization.
}
\description{
\code{hapDendo} creates an hierarchical clustering graphical output of multiple individuals haplotypes.
}
\details{
A \code{data.frame} created by \code{createFullHaplotype}.
}
\examples{
# Load example data and germlines
data(samples_db, HVGERM, HDGERM)

# Infering haplotype
hap_df = createFullHaplotype(samples_db,toHap_col=c("V_CALL","D_CALL"),hapBy_col="J_CALL",hapBy="IGHJ6",toHap_GERM=c(HVGERM,HDGERM));
hapDendo(hap_df)

}
