% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_functions.R
\name{hapHeatmap}
\alias{hapHeatmap}
\title{Graphical output of alleles division by chromosome}
\usage{
hapHeatmap(hap_table, chain = c("IGH", "IGK", "IGL"), ...)
}
\arguments{
\item{hap_table}{haplotype summary table. See details.}

\item{chain}{the IG chain: IGH,IGK,IGL. Default is IGH.}
}
\value{
a single chromosome deletion visualization.
}
\description{
\code{deletionHeatmap} creates a graphical output of the alleles per gene in chromosome in multiple samples.
}
\details{
A \code{data.frame} created by \code{createFullHaplotype}.
}
\examples{
# Load example data and germlines
data(samples_db, HVGERM, HDGERM)

# Infering haplotype
hap_df = createFullHaplotype(samples_db,toHap_col=c("V_CALL","D_CALL"),hapBy_col="J_CALL",hapBy="IGHJ6",toHap_GERM=c(HVGERM,HDGERM));
hapHeatmap(hap_df)

}
