% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{HJGERM}
\alias{HJGERM}
\title{Human IGHJ germlines}
\format{
Values correspond to IMGT nuceltoide sequences.
}
\usage{
HJGERM
}
\description{
A \code{character} vector of all 13 human IGHJ germline gene segment alleles
in IMGT Gene-db release 2018-12-4.
}
\references{
Xochelli \emph{et al}. (2014) Immunoglobulin heavy variable
(IGHV) genes and alleles: new entities, new names and implications for
research and prognostication in chronic lymphocytic leukaemia.
\emph{Immunogenetics}. 67(1):61-6.
}
\keyword{data}
