% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{KVGERM}
\alias{KVGERM}
\alias{LVGERM}
\title{Human IGKV germlines}
\format{
Values correspond to IMGT-gaped nuceltoide sequences (with
nucleotides capitalized and gaps represented by '.').

Values correspond to IMGT-gaped nuceltoide sequences (with
nucleotides capitalized and gaps represented by '.').
}
\usage{
KVGERM

LVGERM
}
\description{
A \code{character} vector of all 342 human IGKV germline gene segment alleles
in IMGT Gene-db release 2019-11-18.

A \code{character} vector of all 342 human IGLV germline gene segment alleles
in IMGT Gene-db release 2019-11-18.
}
\keyword{data}
