% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{LJGERM}
\alias{LJGERM}
\title{Human IGLJ germlines}
\format{
Values correspond to IMGT-gaped nuceltoide sequences (with
nucleotides capitalized and gaps represented by '.').
}
\usage{
LJGERM
}
\description{
A \code{character} vector of all 342 human IGLJ germline gene segment alleles
in IMGT Gene-db release 2019-11-18.
}
\keyword{data}
