% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_functions.R
\name{plotDeletionsByBinom}
\alias{plotDeletionsByBinom}
\title{Graphical output of double chromosome deletions}
\usage{
plotDeletionsByBinom(GENE.usage.df, chain = c("IGH", "IGK", "IGL"), ...)
}
\arguments{
\item{GENE.usage.df}{Double chromosome deletion summary table. See details.}

\item{chain}{the IG chain: IGH,IGK,IGL. Default is IGH.}
}
\value{
a double chromosome deletion visualization.
}
\description{
\code{plotDeletionsByBinom} creates a graphical output of the double chromosome deletions in multiple samples.
}
\details{
A \code{data.frame} created by \code{binom_test_deletion}.
}
\examples{
# Load example data and germlines
data(samples_db)

# Infering haplotype
hap_df = binom_test_deletion(samples_db);
plotDeletionsByBinom(hap_df)

}
