% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_functions.R
\name{plotDeletionsByVpooled}
\alias{plotDeletionsByVpooled}
\title{Graphical output for single chromosome D or J gene deletions according to V pooled method}
\usage{
plotDeletionsByVpooled(del.df, K_ranges = c(3, 7))
}
\arguments{
\item{del.df}{A \code{data.frame} created by \code{deletionsByVpooled}.}

\item{K_ranges}{vector of one or two integers for log(K) certainty level thresholds}
}
\value{
a single chromosome deletion visualization.
}
\description{
\code{plotDeletionsByVpooled} graphical output for single chromosome D or J gene deletions (for heavy chain only).
}
\details{
A \code{data.frame} created by \code{deletionsByVpooled}.
}
\examples{
# Load example data and germlines
data(samples_db)
del_db <- deletionsByVpooled(samples_db)
plotDeletionsByVpooled(del_db)
}
