% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rabhit.R
\docType{package}
\name{rabhit}
\alias{rabhit}
\alias{rabhit-package}
\title{The RAbHIT package}
\description{
The \code{rabhit} package provides a robust novel method for determining
antibody heavy and light chain haplotypes by adapting a Bayesian framework.
The key functions in \code{rabhit}, broken down topic, are
described below.
}
\section{Haplotype and deletion infrence}{

\code{rabhit} provides tools to infer haplotypes based on given anchor genes,
deletion detection based on relative gene usage, pooling v genes, and a single anchor gene.

\itemize{
  \item  \link{createFullHaplotype}:      Infer haplotypes and single chromosome deletions based on an anchor gene.
  \item  \link{deletionsByVpooled}:       Single chromosomal deletion detection by pooling V genes.
  \item  \link{deletionsByBinom}:       Single chromosomal deletion detection by pooling V genes.
}
}

\section{Haplotype visualization}{

Functions for visualization of the inferred haplotypes and deletions

\itemize{
  \item  \link{plotHaplotype}:            Haplotype inference plot.
  \item  \link{deletionHeatmap}:          Single chromosome deletion heatmap.
  \item  \link{hapHeatmap}:               Alleles per genes by chromosome heatmap.
  \item  \link{plotDeletionsByVpooled}:   V pooled based single chromosome deletion heatmap.
  \item  \link{plotDeletionsByBinom}:     Double chromosome deletion heatmap.
}
}

\references{
\enumerate{
  \item  our paper
 }
}
