% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{readHaplotypeDb}
\alias{readHaplotypeDb}
\title{Read a Change-O tab-delimited database file}
\usage{
readHaplotypeDb(file)
}
\arguments{
\item{file}{tab-delimited database file output by a Change-O tool.}
}
\value{
A data.frame of the haplotype file. Columns will be imported as is, except for
          the following columns which will be explicitly converted into character
          values:
          \itemize{
            \item  alleles
            \item  subject
          }
}
\description{
\code{readHaplotypeDb} reads a tab-delimited haplotype file created by a createFullHaplotype
into a data.frame. Based on readChangeoDb function from alakazam.
}
