% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{samples_db}
\alias{samples_db}
\title{Example of IGH human naive b-cell data from multiple
individuals}
\format{A \code{data.frame} in Change-O format (\link{https://changeo.readthedocs.io/en/version-0.4.1---airr-standards/standard.html}) containing the following columns:
\itemize{
  \item \code{"SUBJECT"}: subject names
  \item \code{"V_CALL"}: V allele call(s) (in an IMGT format)
  \item \code{"D_CALL"}: D allele call(s) (in an IMGT format, only for heavy chains)
  \item \code{"J_CALL"}: J allele call(s) (in an IMGT format)
}}
\usage{
samples_db
}
\description{
Example IGH human naive b-cell data from multiple
individuals (see \cite{XXX}).
}
\references{
\cite{XXX}.
}
\keyword{AIRR}
\keyword{NGS}
\keyword{antibody}
\keyword{data}
