% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_functions.R
\name{deletionHeatmap}
\alias{deletionHeatmap}
\title{Graphical output of single chromosome deletions}
\usage{
deletionHeatmap(hap_table, kThreshDel = 3, chain = c("IGH", "IGK",
  "IGL"))
}
\arguments{
\item{hap_table}{haplotype summary table. See details.}

\item{kThreshDel}{the minimum lK (log10 of the Bayes factor) used in \code{createFullHaplotype} to call a deletion. Indicates the color for strong deletion. Defualt is 3.}

\item{chain}{the IG chain: IGH,IGK,IGL. Default is IGH.}
}
\value{
A single chromosome deletion visualization.
}
\description{
The \code{deletionHeatmap} function generates a graphical output of the single chromosome deletions in multiple samples.
}
\details{
A \code{data.frame} created by \code{createFullHaplotype}.
}
\examples{
# Plotting single choromosme deletion from haplotype inference
deletionHeatmap(samplesHaplotype)
}
